support v2.01 GRRCN files with v1.01 back compat
v2.01 was release in October 2019, v3.01 was release in April 2021 but
no support for v3.01 is added here (yet)

from the Amex "Raw Data - Global Reconciliation (GRRCN) Technical File
Specification April 2021":

"v1.01 of the GRRCN file format will be decommissioned as of October 31,
2021, any files sent as v1.01 will be migrated to v2.01 of the GRRCN
file format. All customers using v1.01 of the file format will be
automatically moved to v2.01 on October 31, 2021." so there is some
urgency here as it is likely (at least from our testing) that v2.01 will
break with the current implementation.

the additional field here is FILLER in the TRANSACTION, SUBMISSION, and
TAXRECORD lines. the code has been tweaked to use the VERSION value in
the header and dynamically set the field_map depending on that, so it
retains backwards compatibility. a test and example file has been added
for v2.01 of the GRRCN file